#161 refactor: 폼에 사용되는 인풋들 분리
RecruitmentFormInputs는 어떤 인풋이 필요한지 담당
폼 관련 로직들은 wrapper component가 담당
